build: parallelize travis tests
Unfortunately, our test-suite must wait for timeouts in ACD to expire,
otherwise we cannot test for proper state-engine transitions. To reduce
the test-suite runtime, make travis run the tests in parallel. We do not
care for the number of cores we run on here, but rather just want our
tests to idle in parallel...

Also add a comment to the travis build-script that explains it.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>