n-acd: bail out after raising an event
If we raise an event and thus call into the user-callback, we better
always bail out of the dispatcher. There is no harm in bailing up, since
on the next iteration we just simply continue. However, in case the user
modified the ACD context in the callback, we really want to start over.
Furthermore, we want to give the user a chance to schedule more
important operations, rather then continuing ACD.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>