n-acd: keep context referenced during dispatch
We want to explicitly allow any kind of operation in the ACD callbacks.
Allow the caller to release their last reference, if they want to. This
already works, since we return to the caller after any callback, anyway.

However, this is quite fragile, so take a ref during dispatch to protect
the object from being erased.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>