n-acd: remotable API rework
This is the start of a total rework of the networking APIs, with the aim
of making the APIs "IPC transparent". That firstly means ensuring the
API calls are not inherently synchronous and not unreasonably fine-grained.
The internals will mostly remain the same, but the APIs will be changed as
described below.

* Static configuration

Rather than having a colleciton of individual setters, all configuration
is provided to the engine when calling start(), and in order to change
the configuration the engine must be restarted. The configuration is
given as an open-coded struct.

* Events

Rather than registering callbacks that are triggered from dispatch(),
dispatch() pushes pending events onto a queue, which must be flushed
by calling pop_event() until it runs empty. Each event is represented
as an enum together with an open-coded struct providing data associated
with the event.

* Errors

All negative error codes returned by the library are fatal, and should
not be inspected (apart from as a primitive debugging hint). Positive
error codes are defined and used for anything the client is expected
to handle.

* Getters and setters

Whenever possible these are avoided, as all config should be set as
described above, and the caller should cache information rather than
querying the library. All information is still available to the caller
either because they set it, or because it was given in a callback.

Signed-off-by: Tom Gundersen <teg@jklm.no>