n-acd: rework event API
Firstly, the event structure is now owned by the library, and valid
until the next time an event is popped. Secondly, events are no longer
coalesced, but this may be reintroduced in the future if it has been
determined to be safe.

The only exception to this so far is that stopping the engine flushes
all pending events.

Signed-off-by: Tom Gundersen <teg@jklm.no>