n-acd: introduce a new CONFIGURING state
Between signaling the caller READY and the caller calling announce(),
we cannot defend the address, but at the same time the caller may
rely on not losing the right to use it. Track this as a separate state,
and in this state drop all incoming conflicts. We rely on the following
ANNOUNCE to re-trigger any conflicts and handle that according to
the DEFEND policy.

Signed-off-by: Tom Gundersen <teg@jklm.no>