n-acd: never flush pending events
No longer flush pending events when the engine is stopped, rather require
the caller to pop them all (and either handle them or discard them). The
caller is notified of the engine being stopped by the call to pop_event()
returning N_ACD_E_STOPPED.

It is now an error to restart the engine without firt popping off all
pending events. And it is also an error to call announce() outside of
state CONFIGURING. In both cases N_ACD_E_BUSY is returned.

Signed-off-by: Tom Gundersen <teg@jklm.no>