n-acd: fully stop the engine on USED or CONFLICT
We were stopping the timeouts, but continuing to listen on the socket.
However, in the case of USED or CONFLICT the only thing the user should
do is to stop the engine and potentially restart. This could get awkward
with an asynchronous API (events may be generated after the USED/CONFLICT
event, but before stop() being called).

Now that stop() no longer discards any events, it is safe to always
unconditionally replace the disabling of the timers with a full stop()
to also close the socket and re-enter the INIT state. This means the
caller no longer needs to call stop(), and the fact that they must
call start() again if they want the engine to continue is made more
explicit.

Signed-off-by: Tom Gundersen <teg@jklm.no>