n-acd: make timeout configurable
Allow the maximum number of milliseconds a successfull probe should
take to be configured. The minimum is roughly half, and other timeouts
are adjusted accordingly.

This value should reflect the maximal RTT on the local link, a reasonable
default may be 100ms and in order to comply with the RFC 9000ms should be
used.

Signed-off-by: Tom Gundersen <teg@jklm.no>