eBPF: add eBPF helper functions
This allows maintaining from userspace a set of IP addresses, and to
create an eBPF program bound to such a set and a mac address. The
eBPF only allows through ethernet ARP packets not sent from the given
mac address, and where either the sender or destination IP address is
in the given set.