eBPF: never return packets that userspace should unconditionally drop
Make the filter stricter to not let through packets except the ones
userspace has explicitly requested. Packet delivery is still async, so
userspace still needs to check that they still want the packet at the
time it is received, but this is as strict as we are able to get.

Signed-off-by: Tom Gundersen <teg@jklm.no>