n-acd: rework API to support many probes on a context
This reworks the n-acd API to share kernel resources across multiple
probes. That is, it now supports running an infinite number of probes
on a single NAcd context, while the number of required kernel resources
stays O(1).

This is not ready for release, yet. This needs some more cleanup and
re-hookup of all the tests. Just pushed as a pre-release so we can
start co-working on this.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>