tests: add veth test
This essentially mimics a real nework with two peers.

Run one ACD context on each end of the tunnel. On one end probe for N,
addresses on the other end pre-configure N/3 of the same addresses and probe
for another N/3 of the addresses.

Verify that in the case of simultaneous probes of the same address at most one
succeed, in the case of probing for a configured address it always fails, and
probing for a non-existent address always succeeds.

Make sure to keep N fairly high as the protocol is probabilistic, and we also
want to verify that resizing the internal maps works correctly.

Signed-off-by: Tom Gundersen <teg@jklm.no>