TODO: drop remaining items
On second thoughts we will not implement these items.

Increasing the granularity of timeouts from ms to ns would be more
consistent in general, but would not give us anything in practice,
and when it comes to network APIs ms appears to be the standard.

We could allow any number of retries before giving up defending an
address, instead of the current never/once/always tristate. However,
that would contradict the spec and does not seem to be particularly
useful (but would give us more cases to test).

Signed-off-by: Tom Gundersen <teg@jklm.no>