eBPF: make compile-time optional
The eBPF socket filter is strictly speaking just an optimization, without
it evertyhing works as before, except we are going to be woken up on all
ARP packets, rather than just the ones relevant to us.

Make the eBPF opt-out at compile-time to work also on kernels without eBPF
support.

Signed-off-by: Tom Gundersen <teg@jklm.no>