test-bpf: skip test in case of unsufficient privs
We need sufficient privileges to modify BPF maps. Use the test_setup()
helper as a workaround to verify we have the sufficient privileges. We
don't strictly need a separate network-namespace, but it is the easiest
way to verify this.

This fixes an issue were we cannot run BPF calls in unprivileged docker
containers, since we need capabilities in the root user-ns.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>