util/timer: fall back to CLOCK_MONOTONIC if necessary
There are still supported systems out there that don't provide
CLOCK_BOOTTIME. In those cases, lets fall back to CLOCK_MONOTONIC. Note
that we still require recent kernel headers for compilation, but allow
to be run on older systems.

The effect of this is that we might miss a system-suspend, and after a
wakeup will not see the timer jump. Instead, we think no time passed
and continue our probes. That is, worst case we continue a probe longer
than necessary. This seems like an acceptable tradeoff, so lets support
these older systems.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>